Move ComputeExcisionBoundaryVolQuantities to interpolation
