+++ src/ParallelAlgorithms/Interpolation/CMakeLists.txt
@@ -9,6 +9,8 @@ spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
+  ComputeExcisionBoundaryVolumeQuantities.hpp
+  ComputeExcisionBoundaryVolumeQuantities.tpp
   Interpolate.hpp
   InterpolatedVars.hpp
   InterpolationTarget.hpp
