+++ tests/Unit/ParallelAlgorithms/ApparentHorizonFinder/CMakeLists.txt
@@ -7,7 +7,6 @@ set(LIBRARY_SOURCES
   Test_ApparentHorizonFinder.cpp
   Test_CleanUp.cpp
   Test_ComputeCoords.cpp
-  Test_ComputeExcisionBoundaryVolumeQuantities.cpp
   Test_ComputeHorizonVolumeQuantities.cpp
   Test_ComputeVarsToInterpolateToTarget.cpp
   Test_CurrentTime.cpp
